Create EC50_as_function_of_kdo
This example is with 3 binding sites per RNA. First, the data is generated in one for loop, which goes trough the kdo values and the oligo concentrations and runs Gillespie. For each set of values 5 experiments (100 data points each) are run to find Trel values.
All generated data is saved in separate .txt files for each experiment. The statistics of each experiment is also saved in separate .txt files (for example Trel).
Then, a second for loop loads the Trel values and uses average Trel values of the 5 experiments to find the EC50.
Finally the EC50 values are plottet against the kdo values.
Similar code was used for 1 target site pr RNA and for 2 target sites per RNA.